fix email and phone not showing in org export
